Make palette functions accept 32 bit offsets and sizes Doesn't fix anything but I did it just in case

+++ include/palette.h
@@ -58,9 +58,9 @@ extern u8 ALIGNED(4) gPaletteDecompressionBuffer[];
 extern u16 ALIGNED(4) gPlttBufferUnfaded[PLTT_BUFFER_SIZE];
 extern u16 ALIGNED(4) gPlttBufferFaded[PLTT_BUFFER_SIZE];
 
-void LoadCompressedPalette(const u32 *src, u16 offset, u16 size);
-void LoadPalette(const void *src, u16 offset, u16 size);
-void FillPalette(u16 value, u16 offset, u16 size);
+void LoadCompressedPalette(const u32 *src, u32 offset, u32 size);
+void LoadPalette(const void *src, u32 offset, u32 size);
+void FillPalette(u16 value, u32 offset, u32 size);
 void TransferPlttBuffer(void);
 u8 UpdatePaletteFade(void);
 void ResetPaletteFade(void);
